+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v5.5.0-alpha.4
+
+### 9 May 2023
+
+## v5.5.0-alpha.4
+
+### 9 May 2023
+
 **New**
 
 - Footer: allows customising the `legal_summary`